Final big refactoring
Remove .component and .container file endings and update the README, INSTALLATION and
DOCUMENTATION files.
