+++ src/auth/RequestResetPassword.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link, browserHistory } from "react-router";
 import Validate from "../validate/Validate";
-import ValidateError from "../ui/Error.component";
-import Errors from "../ui/Errors.component";
+import ValidateError from "../ui/Error";
+import Errors from "../ui/Errors";
 
 export class RequestResetPassword extends React.Component {
   constructor() {
