+++ src/notification/NotificationList.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import moment from "moment";
 import Validate from "../validate/Validate";
-import ValidateError from "../ui/Error.component";
+import ValidateError from "../ui/Error";
 
 export class NotificationList extends Component {
